behandlungsjournal List-Detail Patient-Logic moved to List-Detail
