+++ src/app/bhjournal/bhjournal-detail/bhjournal-detail.component.html
@@ -1,22 +1,22 @@
-<div *ngIf="bhjournal">
+<div *ngIf="bhJournal">
   <div class="form-group row">
-    <h2 class="col-xs-7">{{bhjournal.name}}</h2>
+    <h2 class="col-xs-7">{{bhJournal.name}}</h2>
   </div>
   <div class="form-group row">
     <label class="col-xs-3 col-form-label">Name des Behandlungsjournals: </label>
-    <input [(ngModel)]="bhjournal.name" class="col-xs-7" type="text" placeholder="Name"/>
+    <input [(ngModel)]="bhJournal.name" class="col-xs-7" type="text" placeholder="Name"/>
   </div>
   <div class="form-group row">
     <label class="col-xs-3 col-form-label">Diagnose: </label>
-    <input [(ngModel)]="bhjournal.diagnose" class="col-xs-7" type="text" placeholder="Tierart"/>
+    <input [(ngModel)]="bhJournal.diagnose" class="col-xs-7" type="text" placeholder="Tierart"/>
   </div>
 
-  <app-dauer [isEditing]="true" [dauer]="bhjournal.dauer" [labelStartDatum]="labelStart" [labelEndeDatum]="labelEnde"
+  <app-dauer [isEditing]="true" [dauer]="bhJournal.dauer" [labelStartDatum]="labelStart" [labelEndeDatum]="labelEnde"
              (StartDatumChange)="onStartDatumChanged($event)"
              (EndeDatumChange)="onEndeDatumChanged($event)"></app-dauer>
 
   <div class="form-group row">
-    <button class="col-xs-3 col-form-label btn btn-sm btn-primary" type="submit" (click)="saveBhJournal(bhjournal)">
+    <button class="col-xs-3 col-form-label btn btn-sm btn-primary" type="submit" (click)="saveBhJournal(bhJournal)">
       Speichern
     </button>
     <!--<button class="col-xs-7 col-form-label btn btn-sm btn btn-secondary" type="reset" (click)="onCancel()">-->
