+++ src/app/bhjournal/bhjournal-list/bhjournal-list.component.ts
@@ -1,24 +1,14 @@
-import {Component, OnInit, OnDestroy, Output, Input, EventEmitter} from '@angular/core';
-import {Subscription} from 'rxjs';
+import {Component, OnInit, Output, Input, EventEmitter} from '@angular/core';
 import {BhJournal} from '../model/bhjournal';
-import {Patient} from '../../patient/model/patient';
-import {BhJournalService} from '../service/bhjournal.service';
-import {PatientService} from '../../patient/service/patient.service';
-import {MessageService} from '../../shared/service/message/message.service';
-
 
 @Component({
   selector: 'app-bhjournal-list',
   templateUrl: 'bhjournal-list.component.html',
   styleUrls: ['bhjournal-list.component.css']
 })
-export class BhjournalListComponent implements OnInit, OnDestroy {
+export class BhjournalListComponent implements OnInit {
   private isLoading = true;
-  private subscriptionPatient: Subscription;
-  private messageServicePatient: MessageService<Patient>;
-  private selectedPatient: Patient;
   private selectedJournal: BhJournal;
-  private patient_id: string;
 
   /* tslint:disable-next-line:no-unused-variable */
   private labelTherapieStart = 'Beginn der Therapie';
@@ -29,36 +19,8 @@ export class BhjournalListComponent implements OnInit, OnDestroy {
   @Output() bhJournalChange: EventEmitter<BhJournal> = new EventEmitter<BhJournal>();
   @Output() bhJournalCountChange: EventEmitter<number> = new EventEmitter<number>();
 
-  constructor(private bhjournalService: BhJournalService, private patientService: PatientService) {
-    this.messageServicePatient = patientService.messageService;
-    this.subscriptionPatient = this.messageServicePatient.Itemselected$.subscribe(
-      patient => {
-        this.selectedPatient = patient;
-        this.getJournalsbyPatient(this.selectedPatient._id);
-      });
-  }
-
   ngOnInit() {
-    if (typeof (this.selectedPatient) !== 'undefined') {
-      this.patient_id = this.selectedPatient._id;
-    } else {
-      // Component called by path
-      this.selectedPatient = this.patientService.readCache();
-      this.patient_id = this.selectedPatient._id;
-      console.log('Bhjournal-Component ngOnInit selected Patient is undefined');
-    }
-    // Abfrage aller Journale eines Patiente, da das subscribe nur wirkt, wenn die Komponente initialisiert ist
-    // wenn der Patient ausgewählt wird.
-    this.getJournalsbyPatient(this.patient_id);
-
-  }
-
-  private getJournalsbyPatient(patient_id: string) {
-    this.bhjournalService.getJournalsbyPatient_id(patient_id).subscribe(
-      journal => this.getJournals(journal),
-      error => console.log('getJournalsbyPatient', error),
-      () => this.isLoading = false
-    );
+     this.getJournals(this.journals);
   }
 
   private getJournals(journals: Array<BhJournal>) {
@@ -66,6 +28,7 @@ export class BhjournalListComponent implements OnInit, OnDestroy {
     // this.bhJournalChange.emit(journals.length);
     this.journals = journals;
     console.log(this.journals.length);
+    this.isLoading = false;
     // this.bhJournalChange.emit(this.journals.length);
   }
 
@@ -75,8 +38,4 @@ export class BhjournalListComponent implements OnInit, OnDestroy {
     this.bhJournalChange.emit(journal);
   }
 
-  ngOnDestroy() {
-    // prevent memory leak when component destroyed
-    this.subscriptionPatient.unsubscribe();
-  }
 }
